refactor: consolidate documentation by moving MDX files to docs directory
- Moved 18 MDX files from src/pages/ to docs/
- Removed 7 duplicate markdown files
- Updated all TypeScript entry files to import from ../../docs/
- Updated all MDX files to import components from ../src/components/
- Fixed broken .md links in MDX files (changed to .html)
- Build tested and working

Benefits:
- Single source of truth: Documentation exists only in docs/
- No duplication: Eliminated manual sync between markdown and HTML
- Interactive components: MDX files support embedded React components
- Maintainability: Changes only need to be made in one place
