+++ docs/ExamplesIndex.mdx
@@ -1,5 +1,5 @@
-import { Layout } from '../components/Layout';
-import { CodeEditor } from '../components/CodeEditor';
+import { Layout } from '../src/components/Layout';
+import { CodeEditor } from '../src/components/CodeEditor';
 
 <Layout>
