+++ docs/GrammarReference.mdx
@@ -1,5 +1,5 @@
-import { PageLayout } from '../components/PageLayout';
-import { CodeEditor } from '../components/CodeEditor';
+import { PageLayout } from '../src/components/PageLayout';
+import { CodeEditor } from '../src/components/CodeEditor';
 
 <PageLayout title="Grammar Reference">
