+++ docs/QuickStart.mdx
@@ -1,5 +1,5 @@
-import { PageLayout } from '../components/PageLayout';
-import { Carousel } from '../components/Carousel';
+import { PageLayout } from '../src/components/PageLayout';
+import { Carousel } from '../src/components/Carousel';
 
 <PageLayout title="Quick Start">
