+++ docs/SyntaxGuide.mdx
@@ -1,5 +1,5 @@
-import { Layout } from '../components/Layout';
-import { CodeEditor } from '../components/CodeEditor';
+import { Layout } from '../src/components/Layout';
+import { CodeEditor } from '../src/components/CodeEditor';
 
 <Layout>
 
@@ -283,12 +283,12 @@ process "処理";`}
 - Quoted node identifiers (e.g., `"node with spaces"`) are not supported
 - Negative numbers in attributes are not currently supported
 
-See [Testing Approach](testing-approach.md) for validation details.
+See [Testing Approach](testing-approach.html) for validation details.
 
 ## See Also
 
-- [Language Overview](language-overview.md) - Conceptual introduction
-- [Examples Index](examples-index.md) - All examples organized by category
-- [Testing Approach](testing-approach.md) - Validation methodology
+- [Language Overview](language-overview.html) - Conceptual introduction
+- [Examples Index](examples-index.html) - All examples organized by category
+- [Testing Approach](testing-approach.html) - Validation methodology
 
 </Layout>
