+++ docs/VscodeExtension.mdx
@@ -1,4 +1,4 @@
-import { PageLayout } from '../components/PageLayout';
+import { PageLayout } from '../src/components/PageLayout';
 
 <PageLayout title="VS Code Extension">
