--- src/pages/TestingApproach.mdx
@@ -1,249 +1,0 @@
-import { Layout } from '../components/Layout';
-import { CodeEditor } from '../components/CodeEditor';
-
-<Layout>
-
-# DyGram Testing Approach
-
-Comprehensive guide to the generative testing methodology used to validate DyGram.
-
-## Philosophy
-
-Rather than static test cases, DyGram uses **generative testing** to validate language completeness and transformation losslessness. This approach:
-
-- **Generates diverse examples** programmatically instead of hand-writing test cases
-- **Validates actual output** (Mermaid diagrams, JSON) instead of just checking for errors
-- **Discovers edge cases** that static tests miss
-- **Creates inspection artifacts** for manual review
-
-This methodology was developed after discovering that 109 static tests missed critical bugs that 23 generative tests found.
-
-## Test Pipeline
-
-Each test validates the complete transformation pipeline:
-
-```
-DyGram source → Parser → AST → Generator → JSON
-                                        ↓
-                                   Generator → Mermaid
-```
-
-### Validation Steps
-
-1. **Parse Validation**
-   - No parser errors
-   - No lexer errors
-   - AST structure is correct
-
-2. **Completeness Validation**
-   - All nodes from source appear in JSON
-   - All edges from source appear in JSON
-   - No information lost during parsing
-
-3. **Losslessness Validation**
-   - Machine title preserved in Mermaid output
-   - All nodes appear in Mermaid output
-   - Edge labels preserved in Mermaid output
-
-4. **Mermaid Validation**
-   - Output is valid Mermaid syntax
-   - Structural elements present (classDiagram-v2, nodes, edges)
-
-## Example Categories
-
-Tests are organized by language feature:
-
-### Basic Syntax
-- Minimal machines
-- Simple node declarations
-- Typed nodes (task, state, init, context)
-- Node labels
-
-**Examples:** `examples/basic/*.dygram`
-
-### Attributes
-- Typed attributes (string, number, boolean, array)
-- Untyped attributes
-- Complex attribute values
-- Multiple attributes per node
-
-**Examples:** `examples/attributes/*.dygram`
-
-### Edges
-- Basic edges
-- Arrow type variants (→, -->, =>, bidirectional)
-- Edge labels (simple, quoted, with attributes)
-- Chained edges
-
-**Examples:** `examples/edges/*.dygram`
-
-### Nesting
-- Multiple levels of hierarchy
-- Mixed nesting patterns
-- Deep nesting (5+ levels)
-
-**Examples:** `examples/nesting/*.dygram`
-
-### Complex Features
-- Context definitions
-- Workflows
-- Conditional edges
-- Unicode support
-
-**Examples:** `examples/complex/*.dygram`
-
-### Stress Tests
-- Large machines (50+ nodes)
-- Many edges and cross-connections
-- Performance validation
-
-**Examples:** `examples/stress/*.dygram`
-
-### Edge Cases
-- Special characters in identifiers
-- Empty nodes
-- Multiple edges from single node
-- Boundary conditions
-
-**Examples:** `examples/edge-cases/*.dygram`
-
-## Test Implementation
-
-Tests are implemented in `test/integration/generative.test.ts`:
-
-<CodeEditor
-    initialCode={`test('Basic: Minimal Machine', async () => {
-    const source = fs.readFileSync('examples/basic/minimal.dygram', 'utf-8');
-    const result = await runGenerativeTest('Minimal Machine', source);
-    expect(result.passed).toBe(true);
-});`}
-    language="typescript"
-    readOnly
-    height="140px"
-/>
-
-Each test:
-1. Loads source from example file
-2. Parses to AST
-3. Generates JSON and Mermaid
-4. Validates completeness and losslessness
-5. Records results for reporting
-
-## Test Artifacts
-
-Each test run generates artifacts in `test-output/generative/`:
-
-### Individual Test Files
-Each test creates `{test-name}.md` with:
-- Source DyGram code
-- Generated JSON
-- Generated Mermaid diagram
-- Validation results
-
-### Comprehensive Report
-`REPORT.md` contains:
-- Summary statistics (passed/failed/success rate)
-- Issue summary by type
-- Detailed failure analysis
-- All validation errors
-
-## Bugs Found by Generative Testing
-
-### 1. Deep Nesting Bug (CRITICAL)
-**Location:** `src/language/generator/generator.ts:75-92`
-
-**Issue:** JSON generator only handled 2 levels of nesting, losing grandchildren (3+ levels)
-
-**Example:**
-
-<CodeEditor
-    initialCode={`level1 {
-    level2 {
-        level3;  // This was being lost!
-    }
-}`}
-    language="dygram"
-    readOnly
-    height="120px"
-/>
-
-**Fix:** Made `serializeNodes()` recursive to handle arbitrary nesting depth
-
-### 2. Edge Label Loss in Mermaid (IMPORTANT)
-**Location:** `src/language/generator/generator.ts:120-140`
-
-**Issue:** Edge labels captured in JSON but not rendered in Mermaid diagrams
-
-**Example:** `start -init-> middle` was not showing "init" label in output
-
-**Fix:** Added `value` field alongside `attributes` for compatibility
-
-## Test Statistics
-
-Current test coverage:
-- **35 generative tests** covering all language features
-- **100% success rate** after bug fixes
-- **Validates:** Parsing, JSON generation, Mermaid generation, completeness, losslessness
-
-## Running Tests
-
-```bash
-npm test
-```
-
-Test output includes:
-- Real-time validation results
-- Generated artifacts in `test-output/generative/`
-- Comprehensive report with statistics
-
-## Adding New Tests
-
-To add a new test:
-
-1. Create example file in appropriate category:
-
-<CodeEditor
-    initialCode={`echo 'machine "New Feature"\\nnewNode;' > examples/basic/new-feature.dygram`}
-    language="bash"
-    readOnly
-    height="60px"
-/>
-
-2. Add test case in `test/integration/generative.test.ts`:
-
-<CodeEditor
-    initialCode={`test('Basic: New Feature', async () => {
-    const source = fs.readFileSync('examples/basic/new-feature.dygram', 'utf-8');
-    const result = await runGenerativeTest('New Feature', source);
-    expect(result.passed).toBe(true);
-});`}
-    language="typescript"
-    readOnly
-    height="140px"
-/>
-
-3. Run tests and review artifacts:
-
-<CodeEditor
-    initialCode={`npm test
-cat test-output/generative/new_feature.md`}
-    language="bash"
-    readOnly
-    height="80px"
-/>
-
-## Best Practices
-
-1. **Use real examples** - Test cases should represent actual usage patterns
-2. **Validate output** - Always check generated JSON and Mermaid, not just absence of errors
-3. **Create artifacts** - Generate inspection files for manual review
-4. **Test edge cases** - Include boundary conditions and unusual patterns
-5. **Document findings** - Record bugs found and fixes applied
-
-## See Also
-
-- [Syntax Guide](syntax-guide.html) - Language syntax reference
-- [Examples Index](examples-index.html) - All test examples
-- [Language Overview](language-overview.html) - Conceptual introduction
-
-</Layout>
