+++ src/pages/api.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import ApiPage from './Api.mdx';
+import ApiPage from '../../docs/Api.mdx';
 
 const root = document.getElementById('root');
 if (root) {
