+++ src/pages/examples.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import Examples from './Examples.mdx';
+import Examples from '../../docs/Examples.mdx';
 
 const root = document.getElementById('root');
 if (root) {
