+++ src/pages/examples/index.mdx
@@ -1,0 +1,29 @@
+import { Layout } from '../components/Layout';
+
+<Layout>
+
+# Example Categories
+
+Browse DyGram examples organized by feature category:
+
+## [Advanced](advanced.html)
+
+Advanced Features Examples
+
+## [Context](context.html)
+
+Context Management Examples
+
+## [Documentation](documentation.html)
+
+Documentation Features Examples
+
+## [Validation](validation.html)
+
+Validation Features Examples
+
+## [Workflows](workflows.html)
+
+View the generated files
+
+</Layout>
