+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import IndexPage from './Index.mdx';
+import IndexPage from '../../docs/Index.mdx';
 
 const root = document.getElementById('root');
 if (root) {
