+++ src/pages/libraries.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import Libraries from './Libraries.mdx';
+import Libraries from '../../docs/Libraries.mdx';
 
 const root = document.getElementById('root');
 if (root) {
