+++ src/pages/quick-start.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import QuickStart from './QuickStart.mdx';
+import QuickStart from '../../docs/QuickStart.mdx';
 
 const container = document.getElementById('root');
 if (container) {
