+++ src/pages/syntax-guide.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import SyntaxGuide from './SyntaxGuide.mdx';
+import SyntaxGuide from '../../docs/SyntaxGuide.mdx';
 
 const container = document.getElementById('root');
 if (container) {
